… third_party/intel/lib/TritonIntelGPUTransforms/RewriteTensorPointer.cpp