@chengjunlu 2024-09-26
@alexbaden 
I have created an issue to track the performance gap of the transpose and non-transpose B cases.
https://github.com/intel/intel-xpu-backend-for-triton/issues/2354