@alexbaden 2024-09-26
Yes, this change is only to keep rewrite tensor pointer from removing the blocked load.