@etiotto 2024-09-26
I think that removing the `RewriteTensorPointer` pass as in PR #2359 is the way to go. We want to avoid rewriting block pointers into regular pointers to maximize the number of `tt.load/tt.store` operations that can be lowered to 2D block/read/write instructions, so removing that pass seems the right approach to me. 

If we land PR #2359 this PR can be rebased presumably and the LIT test kept.